+++ NEWS.md
@@ -6,6 +6,7 @@
 
 * Reloading models saved to file working again
 * New function `update_saved_state_barcode_plaf` to update plaf in saved simulation
+* New function `update_saved_state_barcode_haplotype` to update haplotypes in saved simulation
 
 ## magenta 1.3.5